Rename profile.py
